api/requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pyyaml==6.0.1
python-multipart==0.0.6